iommu: make counter long long
A regular unsigned int may not be enough to track the number of
outstanding ephemeral mappings an may wrap.

long long is not at risk of wrapping since we can have at most
UINT64_MAX / PAGESIZE mappings.

+++ src/vfio/iommu.h
@@ -28,7 +28,7 @@ struct iommu_state {
 	pthread_mutex_t lock;
 
 	uint64_t top, bottom;
-	unsigned int nephemeral;
+	unsigned long long nephemeral;
 
 	void *map;
 };
